Winston Lutz QA test program in Matlab
Run "W_L_test.m" .  Using the file menu open a Winston Lutz test scan. a
single sample is included in this directory "07202012.tif" .  In the new
window crop the image. try to exclude the parts written by the marker.
Next, roughly click on the circles. after clicking on the last one do a
left click. The program finds the centers of all ten circles.